+++ me/scene/SceneManager.h
@@ -36,7 +36,7 @@ namespace me::scene
 		size_t GetSceneCount() const;
 		void AddScene( IScene::ptr scene );
 
-		int FindSceneIndex( std::string name );
+		size_t FindSceneIndex( std::string name );
 		std::string GetSceneName( int index );
 
 		IScene* GetCurrentScene();
